backend/pyproject.toml
[project]
name = "textbook-qa"
version = "1.0.0"
description = "RAG-based Educational Question Answering System with Deep Source Tracing"
requires-python = ">=3.10"

dependencies = [
    "ollama>=0.4.0",
    "chromadb>=0.5.0",
    "sentence-transformers>=3.0.0",
    "streamlit>=1.40.0",
    "PyMuPDF>=1.24.0",
    "Pillow>=10.0.0",
    "pyyaml>=6.0",
    "loguru>=0.7.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0.0",
    "ruff>=0.8.0",
]
